backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
groq
python-multipart
openpyxl
